GPU Init fix and textures
Some more cleanup
Moved GPU init to databuf
Use Short instead of Integer
Rename thing from cpuPseudoProcessor to cpu6502Processor
Change UnlocalizedName to cpu6502
Put it in the OC tab
Give it an texture
don't register items under "oc:cpu_pseudo_processor"

+++ src/main/java/gamax92/ocsymon/SymonArchitecture.java
@@ -93,7 +93,7 @@ public class SymonArchitecture implements Architecture {
 					signal = machine.popSignal();
 					if (signal != null) {
 						if (signal.name().equals("key_down")) {
-							int character = (int)(double)(Double)signal.args()[1]; // castception
+							short character = (short)(double)(Double)signal.args()[1]; // castception
 							if (character != 0) // Not a character
 								console.pushChar(character);
 						}
